use bigvgan-base as default; downloader from gdrive
